No remote file retrieval in DrupalFiles class.

+++ modules/common/tests/src/Unit/Util/DrupalFilesTest.php
@@ -16,10 +16,7 @@ use Symfony\Component\DependencyInjection\ContainerInterface;
  */
 class DrupalFilesTest extends TestCase {
 
-  /**
-   *
-   */
-  public function test() {
+  public function testLocalFile() {
     $drupalFiles = DrupalFiles::create($this->getContainer());
     $drupalFiles->retrieveFile(
       "file://" . __DIR__ . "/../../../files/hello.txt",
@@ -27,9 +24,36 @@ class DrupalFilesTest extends TestCase {
     $this->assertTrue(file_exists("/tmp/hello.txt"));
   }
 
+  public function testBadScheme() {
+    $drupalFiles = DrupalFiles::create($this->getContainer());
+    $this->expectExceptionMessage("Only file:// and http(s) urls are supported");
+    $drupalFiles->retrieveFile(
+      "public://hello.txt",
+      "public://tmp");
+    $this->assertTrue(file_exists("/tmp/hello.txt"));
+  }
+
+  public function testBadDestination() {
+    $drupalFiles = DrupalFiles::create($this->getContainer());
+    $this->expectExceptionMessage("Only moving files to Drupal's public directory (public://) is supported");
+    $drupalFiles->retrieveFile(
+      "file://hello.txt",
+      "file://tmp");
+    $this->assertTrue(file_exists("/tmp/hello.txt"));
+  }
+
   /**
-   * Private.
+   * Added after system_retrieve_file was deprecated.
    */
+  public function testRemoteFile() {
+    $drupalFiles = DrupalFiles::create($this->getContainer());
+    $this->expectExceptionMessage("Remote file retrieval not yet supported");
+    $drupalFiles->retrieveFile(
+      "https://web/hello.txt",
+      "public://tmp/hello.txt");
+    $this->assertTrue(file_exists("/tmp/hello.txt"));
+  }
+
   private function getContainer(): ContainerInterface {
     $options = (new Options())
       ->add('file_system', FileSystemInterface::class)
@@ -44,12 +68,11 @@ class DrupalFilesTest extends TestCase {
       ->getMock();
   }
 
-  /**
-   * Protected.
-   */
   protected function tearDown(): void {
     parent::tearDown();
-    unlink("/tmp/hello.txt");
+    if (file_exists("/tmp/hello.txt")) {
+      unlink("/tmp/hello.txt");
+    }
   }
 
 }
